Script to count outliers.
Outliers are assumed to be data points greater than a given percentile.
For any combination of min--max percentiles and min--max sliding window sizes, this script calculates the number of data which would be excluded by each possible choice of window size / percentile.

The script generates a JSON file containing outliers counts in this format:

{
  'TOTAL_ITERATIONS' : <int>,
  '1' : '1' : 1000,   # window size -> percentile -> number of outliers
   ...
}

The script also generates a LaTeX file with the data presented as a table, and a PDF with the data presented as a 3d plot.
Because this script is slow, it will run in parallel in as many processes as the machine has CPUs.