Allow user to plot specific executions from different benchmarks and Krun files.
The -b CLI switch can now be used repeatedly, in the format:

    -b machine:benchmark:vm:variant:p_exec

e.g.

    -b bencher3:fannkuch_redux:Hotspot:default-java:1 -b bencher5:fasta:PyPy:default-python:0

The get_data_dictionaries() function now collects ONLY the data that the user has requested
(including outliers, if necessary). It also constructs all subplot titles. get_data_dictionaries()
skips any executions or benchmarks which have crashed or been skipped. It checks that the
data gathered from Krun results files is exactly the data requested on the command line.

The main() function now reassembles data from get_data_dictionaries() into "pages" which
are to be plotted. The user can request that all subplots are placed on a single page with
the --one-page switch. Otherwise, each benchmark key for each machine is plotted on a single
page.