Check for (some) packages needed for installation to complete.
This is certainly not a complete list, but I noticed earlier that if these
things are missing, then installation fails in weird or silent fashion (e.g. it
appears to continue successfully, but doesn't install things, and scrolling back
through the terminal one eventually spots the source of the error). It would be
good to think more carefully about what other things are needed, but this is a
start at least.