Allow the user to diff Krun results files. This script is based on the code that generated the charts in Figures 9/11 of the warmup paper. The output is a LaTeX table, similar to the one produced by table_classification_summaries_others, which shows the results of the second input file, and some background colours to show which parts of the table have improved / worsened, etc.
The background colour of the benchmark name shows the 'intersection' of the
other measures for that benchmark. Where NSS's appear in the summaries, we
sometimes cannot compare before/after results, or sometimes we can only say
that the two results differ.

Where a measure for a benchmark has changed, we show the absolute
difference in median / mean above the IQR / CI.

As with some of the other scripts that call the (slow) code to produce
summary information, we dump out a human-readable JSON file containing the
summaries and diff information. This can then be loaded with the -s option
(instead of providing Krun results files directly).