Only say that CIs have narrowed if it exceeds a certain threshold.
In some cases we have CIs narrowing by tiny amounts, which isn't really enough
to be considered "better". This diff sets a minimum threshold of 0.0001s as
being considered significant (note that is is much smaller than the threshold
used to determine if two changepoint segments are correct, since we expect CIs
to be much smaller than means).