Document that it's safe to use imprecise summing
The comment is based on his following experiment and observations:

```
import math, random

x = [random.random() for _ in xrange(100000)]
print("%.10f" % math.fsum(x))
xsum = 0.0
for y in x:
    xsum += y
print("%.10f" % xsum)
```

This prints numbers like:

```
50007.3735223347
50007.3735223343
```

and:

```
50023.2727688112
50023.2727688109
```

and so on. In practice, the error seems at worst to be in the 9th digit for 100,000 values between 0 and 1, which we can happily live with.

Signed-off-by: Stefan Marr <git@stefan-marr.de>