Add a --quality switch to warmup_stats.
PR #97 showed that we can get meaningful performance changes in table generation
if we're prepared to tolerate slightly lower quality stats. In that PR, I made a
rough estimate of how inaccurate they might be. However, I neglected to consider
that floating point numbers have a very uneven spread: 50% of FP numbers reside
between 0 and 1. In other words, the error is going to increase as someone's
benchmark timings increase, which is surprising.

This commit introduces a --quality option with two settings: "low" (and fast)
and "high" (and slow). The default is "high". Relative to #97, "low" also
reduces the bootstrapping iterations, as well as using less accurate floating
point summing. Roughly speaking, "high" is about 3x slower than "low", but one
can observe small changes (at the fifth decimal place onwards) in data between
low and high. Most of the time this is probably OK, but for paper quality
benchmarking, the slightly longer time is worth it for more accurate stats.